Creation of colorizer.py
