+++ lib/matplotlib/colorizer.py
@@ -1,0 +1,502 @@
+"""
+The Colorizer class which handles the data to color pipeline via a
+normalization and a colormap.
+
+.. seealso::
+
+  :doc:`/gallery/color/colormap_reference` for a list of builtin colormaps.
+
+  :ref:`colormap-manipulation` for examples of how to make
+  colormaps.
+
+  :ref:`colormaps` an in-depth discussion of choosing
+  colormaps.
+
+  :ref:`colormapnorms` for more details about data normalization.
+"""
+
+import numpy as np
+from numpy import ma
+import functools
+from matplotlib import _api, colors, cbook, scale, cm
+
+
+class Colorizer():
+    """
+    Class that holds the data to color pipeline
+    accessible via `.to_rgba(A)` and executed via
+    the `.norm` and `.cmap` attributes.
+    """
+    def __init__(self, cmap=None, norm=None):
+
+        self._cmap = None
+        self._set_cmap(cmap)
+
+        self._id_norm = None
+        self._norm = None
+        self.norm = norm
+
+        self.callbacks = cbook.CallbackRegistry(signals=["changed"])
+        self.colorbar = None
+
+    def _scale_norm(self, norm, vmin, vmax, A):
+        """
+        Helper for initial scaling.
+
+        Used by public functions that create a ScalarMappable and support
+        parameters *vmin*, *vmax* and *norm*. This makes sure that a *norm*
+        will take precedence over *vmin*, *vmax*.
+
+        Note that this method does not set the norm.
+        """
+        if vmin is not None or vmax is not None:
+            self.set_clim(vmin, vmax)
+            if isinstance(norm, colors.Normalize):
+                raise ValueError(
+                    "Passing a Normalize instance simultaneously with "
+                    "vmin/vmax is not supported.  Please pass vmin/vmax "
+                    "directly to the norm when creating it.")
+
+        # always resolve the autoscaling so we have concrete limits
+        # rather than deferring to draw time.
+        self.autoscale_None(A)
+
+    @property
+    def norm(self):
+        return self._norm
+
+    @norm.setter
+    def norm(self, norm):
+        _api.check_isinstance((colors.Normalize, str, None), norm=norm)
+        if norm is None:
+            norm = colors.Normalize()
+        elif isinstance(norm, str):
+            try:
+                scale_cls = scale._scale_mapping[norm]
+            except KeyError:
+                raise ValueError(
+                    "Invalid norm str name; the following values are "
+                    f"supported: {', '.join(scale._scale_mapping)}"
+                ) from None
+            norm = _auto_norm_from_scale(scale_cls)()
+
+        if norm is self.norm:
+            # We aren't updating anything
+            return
+
+        in_init = self.norm is None
+        # Remove the current callback and connect to the new one
+        if not in_init:
+            self.norm.callbacks.disconnect(self._id_norm)
+        self._norm = norm
+        self._id_norm = self.norm.callbacks.connect('changed',
+                                                    self.changed)
+        if not in_init:
+            self.changed()
+
+    def to_rgba(self, x, alpha=None, bytes=False, norm=True):
+        """
+        Return a normalized RGBA array corresponding to *x*.
+
+        In the normal case, *x* is a 1D or 2D sequence of scalars, and
+        the corresponding `~numpy.ndarray` of RGBA values will be returned,
+        based on the norm and colormap set for this Colorizer.
+
+        There is one special case, for handling images that are already
+        RGB or RGBA, such as might have been read from an image file.
+        If *x* is an `~numpy.ndarray` with 3 dimensions,
+        and the last dimension is either 3 or 4, then it will be
+        treated as an RGB or RGBA array, and no mapping will be done.
+        The array can be `~numpy.uint8`, or it can be floats with
+        values in the 0-1 range; otherwise a ValueError will be raised.
+        Any NaNs or masked elements will be set to 0 alpha.
+        If the last dimension is 3, the *alpha* kwarg (defaulting to 1)
+        will be used to fill in the transparency.  If the last dimension
+        is 4, the *alpha* kwarg is ignored; it does not
+        replace the preexisting alpha.  A ValueError will be raised
+        if the third dimension is other than 3 or 4.
+
+        In either case, if *bytes* is *False* (default), the RGBA
+        array will be floats in the 0-1 range; if it is *True*,
+        the returned RGBA array will be `~numpy.uint8` in the 0 to 255 range.
+
+        If norm is False, no normalization of the input data is
+        performed, and it is assumed to be in the range (0-1).
+
+        """
+        # First check for special case, image input:
+        # First check for special case, image input:
+        try:
+            if x.ndim == 3:
+                if x.shape[2] == 3:
+                    if alpha is None:
+                        alpha = 1
+                    if x.dtype == np.uint8:
+                        alpha = np.uint8(alpha * 255)
+                    m, n = x.shape[:2]
+                    xx = np.empty(shape=(m, n, 4), dtype=x.dtype)
+                    xx[:, :, :3] = x
+                    xx[:, :, 3] = alpha
+                elif x.shape[2] == 4:
+                    xx = x
+                else:
+                    raise ValueError("Third dimension must be 3 or 4")
+                if xx.dtype.kind == 'f':
+                    # If any of R, G, B, or A is nan, set to 0
+                    if np.any(nans := np.isnan(x)):
+                        if x.shape[2] == 4:
+                            xx = xx.copy()
+                        xx[np.any(nans, axis=2), :] = 0
+
+                    if norm and (xx.max() > 1 or xx.min() < 0):
+                        raise ValueError("Floating point image RGB values "
+                                         "must be in the 0..1 range.")
+                    if bytes:
+                        xx = (xx * 255).astype(np.uint8)
+                elif xx.dtype == np.uint8:
+                    if not bytes:
+                        xx = xx.astype(np.float32) / 255
+                else:
+                    raise ValueError("Image RGB array must be uint8 or "
+                                     "floating point; found %s" % xx.dtype)
+                # Account for any masked entries in the original array
+                # If any of R, G, B, or A are masked for an entry, we set alpha to 0
+                if np.ma.is_masked(x):
+                    xx[np.any(np.ma.getmaskarray(x), axis=2), 3] = 0
+                return xx
+        except AttributeError:
+            # e.g., x is not an ndarray; so try mapping it
+            pass
+
+        # This is the normal case, mapping a scalar array:
+        x = ma.asarray(x)
+        if norm:
+            x = self.norm(x)
+        rgba = self.cmap(x, alpha=alpha, bytes=bytes)
+        return rgba
+
+    def normalize(self, x):
+        """
+        Normalize the data in x.
+
+        Parameters
+        ----------
+        x : np.array
+
+        Returns
+        -------
+        np.array
+
+        """
+        return self.norm(x)
+
+    def autoscale(self, A):
+        """
+        Autoscale the scalar limits on the norm instance using the
+        current array
+        """
+        if A is None:
+            raise TypeError('You must first set_array for mappable')
+        # If the norm's limits are updated self.changed() will be called
+        # through the callbacks attached to the norm
+        self.norm.autoscale(A)
+
+    def autoscale_None(self, A):
+        """
+        Autoscale the scalar limits on the norm instance using the
+        current array, changing only limits that are None
+        """
+        if A is None:
+            raise TypeError('You must first set_array for mappable')
+        # If the norm's limits are updated self.changed() will be called
+        # through the callbacks attached to the norm
+        self.norm.autoscale_None(A)
+
+    def _set_cmap(self, cmap):
+        """
+        Set the colormap for luminance data.
+
+        Parameters
+        ----------
+        cmap : `.Colormap` or str or None
+        """
+        in_init = self._cmap is None
+        cmap = cm._ensure_cmap(cmap)
+        self._cmap = cmap
+        if not in_init:
+            self.changed()  # Things are not set up properly yet.
+
+    @property
+    def cmap(self):
+        return self._cmap
+
+    @cmap.setter
+    def cmap(self, cmap):
+        self._set_cmap(cmap)
+
+    def set_clim(self, vmin=None, vmax=None):
+        """
+        Set the norm limits for image scaling.
+
+        Parameters
+        ----------
+        vmin, vmax : float
+             The limits.
+
+             The limits may also be passed as a tuple (*vmin*, *vmax*) as a
+             single positional argument.
+
+             .. ACCEPTS: (vmin: float, vmax: float)
+        """
+        # If the norm's limits are updated self.changed() will be called
+        # through the callbacks attached to the norm
+        if vmax is None:
+            try:
+                vmin, vmax = vmin
+            except (TypeError, ValueError):
+                pass
+        if vmin is not None:
+            self.norm.vmin = colors._sanitize_extrema(vmin)
+        if vmax is not None:
+            self.norm.vmax = colors._sanitize_extrema(vmax)
+
+    def get_clim(self):
+        """
+        Return the values (min, max) that are mapped to the colormap limits.
+        """
+        return self.norm.vmin, self.norm.vmax
+
+    def changed(self):
+        """
+        Call this whenever the mappable is changed to notify all the
+        callbackSM listeners to the 'changed' signal.
+        """
+        self.callbacks.process('changed')
+        self.stale = True
+
+    @property
+    def vmin(self):
+        return self.get_clim[0]
+
+    @vmin.setter
+    def vmin(self, vmin):
+        self.set_clim(vmin=vmin)
+
+    @property
+    def vmax(self):
+        return self.get_clim[1]
+
+    @vmax.setter
+    def vmax(self, vmax):
+        self.set_clim(vmax=vmax)
+
+    @property
+    def clip(self):
+        return self.norm.clip
+
+    @clip.setter
+    def clip(self, clip):
+        self.norm.clip = clip
+
+
+class ColorizerShim:
+
+    def _scale_norm(self, norm, vmin, vmax):
+        self.colorizer._scale_norm(norm, vmin, vmax, self._A)
+
+    def to_rgba(self, x, alpha=None, bytes=False, norm=True):
+        """
+        Return a normalized RGBA array corresponding to *x*.
+
+        In the normal case, *x* is a 1D or 2D sequence of scalars, and
+        the corresponding `~numpy.ndarray` of RGBA values will be returned,
+        based on the norm and colormap set for this Colorizer.
+
+        There is one special case, for handling images that are already
+        RGB or RGBA, such as might have been read from an image file.
+        If *x* is an `~numpy.ndarray` with 3 dimensions,
+        and the last dimension is either 3 or 4, then it will be
+        treated as an RGB or RGBA array, and no mapping will be done.
+        The array can be `~numpy.uint8`, or it can be floats with
+        values in the 0-1 range; otherwise a ValueError will be raised.
+        Any NaNs or masked elements will be set to 0 alpha.
+        If the last dimension is 3, the *alpha* kwarg (defaulting to 1)
+        will be used to fill in the transparency.  If the last dimension
+        is 4, the *alpha* kwarg is ignored; it does not
+        replace the preexisting alpha.  A ValueError will be raised
+        if the third dimension is other than 3 or 4.
+
+        In either case, if *bytes* is *False* (default), the RGBA
+        array will be floats in the 0-1 range; if it is *True*,
+        the returned RGBA array will be `~numpy.uint8` in the 0 to 255 range.
+
+        If norm is False, no normalization of the input data is
+        performed, and it is assumed to be in the range (0-1).
+
+        """
+        return self.colorizer.to_rgba(x, alpha=alpha, bytes=bytes, norm=norm)
+
+    def get_clim(self):
+        """
+        Return the values (min, max) that are mapped to the colormap limits.
+        """
+        return self.colorizer.get_clim()
+
+    def set_clim(self, vmin=None, vmax=None):
+        """
+        Set the norm limits for image scaling.
+
+        Parameters
+        ----------
+        vmin, vmax : float
+             The limits.
+
+             For scalar data, the limits may also be passed as a
+             tuple (*vmin*, *vmax*) as a single positional argument.
+
+             .. ACCEPTS: (vmin: float, vmax: float)
+        """
+        # If the norm's limits are updated self.changed() will be called
+        # through the callbacks attached to the norm
+        self.colorizer.set_clim(vmin, vmax)
+
+    def get_alpha(self):
+        """
+        Returns
+        -------
+        float
+            Always returns 1.
+        """
+        # This method is intended to be overridden by Artist sub-classes
+        return 1.
+
+    @property
+    def cmap(self):
+        return self.colorizer.cmap
+
+    @cmap.setter
+    def cmap(self, cmap):
+        self.colorizer.cmap = cmap
+
+    def get_cmap(self):
+        """Return the `.Colormap` instance."""
+        return self.colorizer.cmap
+
+    def set_cmap(self, cmap):
+        """
+        Set the colormap for luminance data.
+
+        Parameters
+        ----------
+        cmap : `.Colormap` or str or None
+        """
+        self.cmap = cmap
+
+    @property
+    def norm(self):
+        return self.colorizer.norm
+
+    @norm.setter
+    def norm(self, norm):
+        self.colorizer.norm = norm
+
+    def set_norm(self, norm):
+        """
+        Set the normalization instance.
+
+        Parameters
+        ----------
+        norm : `.Normalize` or str or None
+
+        Notes
+        -----
+        If there are any colorbars using the mappable for this norm, setting
+        the norm of the mappable will reset the norm, locator, and formatters
+        on the colorbar to default.
+        """
+        self.norm = norm
+
+    def autoscale(self):
+        """
+        Autoscale the scalar limits on the norm instance using the
+        current array
+        """
+        self.colorizer.autoscale(self._A)
+
+    def autoscale_None(self):
+        """
+        Autoscale the scalar limits on the norm instance using the
+        current array, changing only limits that are None
+        """
+        self.colorizer.autoscale_None(self._A)
+
+    @property
+    def colorbar(self):
+        return self.colorizer.colorbar
+
+    @colorbar.setter
+    def colorbar(self, colorbar):
+        self.colorizer.colorbar = colorbar
+
+    def _format_cursor_data_override(self, data):
+        # This function overwrites Artist.format_cursor_data(). We cannot
+        # implement cm.ScalarMappable.format_cursor_data() directly, because
+        # most cm.ScalarMappable subclasses inherit from Artist first and from
+        # cm.ScalarMappable second, so Artist.format_cursor_data would always
+        # have precedence over cm.ScalarMappable.format_cursor_data.
+
+        # Note if cm.ScalarMappable is depreciated, this functionality should be 
+        # implemented as format_cursor_data() on ColorizingArtist. 
+        n = self.cmap.N
+        if np.ma.getmask(data):
+            return "[]"
+        normed = self.norm(data)
+        if np.isfinite(normed):
+            if isinstance(self.norm, colors.BoundaryNorm):
+                # not an invertible normalization mapping
+                cur_idx = np.argmin(np.abs(self.norm.boundaries - data))
+                neigh_idx = max(0, cur_idx - 1)
+                # use max diff to prevent delta == 0
+                delta = np.diff(
+                    self.norm.boundaries[neigh_idx:cur_idx + 2]
+                ).max()
+            elif self.norm.vmin == self.norm.vmax:
+                # singular norms, use delta of 10% of only value
+                delta = np.abs(self.norm.vmin * .1)
+            else:
+                # Midpoints of neighboring color intervals.
+                neighbors = self.norm.inverse(
+                    (int(normed * n) + np.array([0, 1])) / n)
+                delta = abs(neighbors - data).max()
+            g_sig_digits = cbook._g_sig_digits(data, delta)
+        else:
+            g_sig_digits = 3  # Consistent with default below.
+        return f"[{data:-#.{g_sig_digits}g}]"
+
+
+def _auto_norm_from_scale(scale_cls):
+    """
+    Automatically generate a norm class from *scale_cls*.
+
+    This differs from `.colors.make_norm_from_scale` in the following points:
+
+    - This function is not a class decorator, but directly returns a norm class
+      (as if decorating `.Normalize`).
+    - The scale is automatically constructed with ``nonpositive="mask"``, if it
+      supports such a parameter, to work around the difference in defaults
+      between standard scales (which use "clip") and norms (which use "mask").
+
+    Note that ``make_norm_from_scale`` caches the generated norm classes
+    (not the instances) and reuses them for later calls.  For example,
+    ``type(_auto_norm_from_scale("log")) == LogNorm``.
+    """
+    # Actually try to construct an instance, to verify whether
+    # ``nonpositive="mask"`` is supported.
+    try:
+        norm = colors.make_norm_from_scale(
+            functools.partial(scale_cls, nonpositive="mask"))(
+            colors.Normalize)()
+    except TypeError:
+        norm = colors.make_norm_from_scale(scale_cls)(
+            colors.Normalize)()
+    return type(norm)
