+++ lib/matplotlib/meson.build
@@ -33,6 +33,7 @@ python_sources = [
   'cm.py',
   'collections.py',
   'colorbar.py',
+  'colorizer.py',
   'colors.py',
   'container.py',
   'contour.py',
